fix(cli): satisfy private candidate policy gate
