+++ private/hosted-cli-candidate/default-services.js
@@ -2,14 +2,11 @@
 
 const crypto = require('node:crypto');
 const path = require('node:path');
-const {
-  configureTargetSetup,
-  checkCredentialSources,
-  readInstallCredentials,
-} = require('./credentials');
+const { checkCredentialSources, readInstallCredentials } = require('./credentials');
 const { SealedInstallClient } = require('./install-client');
 const { HostedRunOrchestrator } = require('./orchestrator');
 const { readHostedInputs } = require('./readers');
+const { createTargetServices, targetSessionManager } = require('./target-services');
 
 function runtimeModule(relative) {
   return require(path.join(__dirname, '..', relative));
@@ -52,29 +49,15 @@ async function createSessionContext(name, runtime, settings) {
   const http = httpTransport();
   const descriptor = await runtime.target.discoverTarget(target.url, http);
   const credentialStore = await runtime.target.KeyringCredentialStore.create();
-  const sessionManager = new runtime.target.TargetSessionManager({
-    targetName: name,
-    target,
-    credentialStore,
-    acquireLock: () => runtime.target.acquireTargetLock(target.id),
+  const sessionManager = targetSessionManager({
+    runtime,
     settings,
-    deps: {
-      http,
-      clock: Date,
-      browserOpener: { open: () => Promise.resolve() },
-      stderr: process.stderr,
-      discoveryEndpoints: {
-        deviceAuthorizationEndpoint: descriptor.oauth.deviceAuthorizationEndpoint,
-        tokenEndpoint: descriptor.oauth.tokenEndpoint,
-        revocationEndpoint: descriptor.oauth.revocationEndpoint,
-        clientId: descriptor.oauth.clientId,
-        capsuleApiBaseUrl: descriptor.capsule.baseUrl.replace(/\/$/, ''),
-        deviceGrantType: descriptor.oauth.deviceGrantType,
-        audience: descriptor.oauth.audience,
-        sessionEndpoint: new URL(descriptor.session.routeTemplate.template, descriptor.origin).href,
-        descriptor,
-      },
-    },
+    name,
+    target,
+    descriptor,
+    credentialStore,
+    open: () => Promise.resolve(),
+    http,
   });
   const adapter = runtime.hostedTarget.createTargetAdapter({
     descriptor,
@@ -108,132 +91,7 @@ function createDefaultServices(dependencies) {
   const runtime = loadRuntime();
   const settings = targetSettings(dependencies);
   const services = {
-    async targetAdd(name, options) {
-      const url = runtime.target.normalizeAndValidateUrl(options.url);
-      const descriptor = await runtime.target.discoverTarget(url, httpTransport());
-      const record = runtime.target.addTarget(name, url, settings, descriptor);
-      console.log(`Target "${name}" added (${record.url})`);
-    },
-
-    async targetLogin(name) {
-      const target = requireTarget(name, runtime, settings);
-      const descriptor = await runtime.target.discoverTarget(target.url, httpTransport());
-      const credentialStore = await runtime.target.KeyringCredentialStore.create();
-      const manager = new runtime.target.TargetSessionManager({
-        targetName: name,
-        target,
-        credentialStore,
-        acquireLock: () => runtime.target.acquireTargetLock(target.id),
-        settings,
-        deps: {
-          http: httpTransport(),
-          clock: Date,
-          browserOpener: {
-            async open(url) {
-              const imported = await import('open');
-              await imported.default(url);
-            },
-          },
-          stderr: process.stderr,
-          discoveryEndpoints: {
-            deviceAuthorizationEndpoint: descriptor.oauth.deviceAuthorizationEndpoint,
-            tokenEndpoint: descriptor.oauth.tokenEndpoint,
-            revocationEndpoint: descriptor.oauth.revocationEndpoint,
-            clientId: descriptor.oauth.clientId,
-            capsuleApiBaseUrl: descriptor.capsule.baseUrl.replace(/\/$/, ''),
-            deviceGrantType: descriptor.oauth.deviceGrantType,
-            audience: descriptor.oauth.audience,
-            sessionEndpoint: new URL(descriptor.session.routeTemplate.template, descriptor.origin)
-              .href,
-            descriptor,
-          },
-        },
-      });
-      const result = await manager.login();
-      console.log(`Logged in to "${name}" (organization: ${result.organization.id})`);
-    },
-
-    async targetList(options) {
-      const targets = runtime.target.listTargets(settings);
-      if (options.json) {
-        const rows = targets.map(({ name, record }) => ({
-          name,
-          id: record.id,
-          url: record.url,
-          organization: record.organization ?? null,
-          configured: record.hostedSetup?.kind === 'zeroshot.private-hosted-setup/v1',
-          createdAt: record.createdAt,
-        }));
-        console.log(JSON.stringify(rows, null, 2));
-        return;
-      }
-      if (targets.length === 0) {
-        console.log('No targets registered.');
-        return;
-      }
-      for (const { name, record } of targets) {
-        console.log(`${name}\t${record.url}\t${record.organization?.id ?? 'not-logged-in'}`);
-      }
-    },
-
-    async targetRemove(name, options) {
-      const target = requireTarget(name, runtime, settings);
-      try {
-        const descriptor = await runtime.target.discoverTarget(target.url, httpTransport());
-        const credentialStore = await runtime.target.KeyringCredentialStore.create();
-        const manager = new runtime.target.TargetSessionManager({
-          targetName: name,
-          target,
-          credentialStore,
-          acquireLock: () => runtime.target.acquireTargetLock(target.id),
-          settings,
-          deps: {
-            http: httpTransport(),
-            clock: Date,
-            browserOpener: { open: () => Promise.resolve() },
-            stderr: process.stderr,
-            discoveryEndpoints: {
-              deviceAuthorizationEndpoint: descriptor.oauth.deviceAuthorizationEndpoint,
-              tokenEndpoint: descriptor.oauth.tokenEndpoint,
-              revocationEndpoint: descriptor.oauth.revocationEndpoint,
-              clientId: descriptor.oauth.clientId,
-              capsuleApiBaseUrl: descriptor.capsule.baseUrl.replace(/\/$/, ''),
-              deviceGrantType: descriptor.oauth.deviceGrantType,
-              audience: descriptor.oauth.audience,
-              sessionEndpoint: new URL(descriptor.session.routeTemplate.template, descriptor.origin)
-                .href,
-              descriptor,
-            },
-          },
-        });
-        await manager.revoke(Boolean(options.force));
-        const setup = target.hostedSetup;
-        if (setup?.openrouter?.service && setup?.openrouter?.account) {
-          await credentialStore.delete(setup.openrouter.service, setup.openrouter.account);
-        }
-      } catch (error) {
-        if (!options.force) throw error;
-      }
-      runtime.target.removeTarget(name, settings);
-      console.log(`Target "${name}" removed`);
-    },
-
-    async targetSetup(name, options) {
-      const target = requireTarget(name, runtime, settings);
-      const credentialStore = await runtime.target.KeyringCredentialStore.create();
-      const metadata = await configureTargetSetup({
-        targetName: name,
-        target,
-        repository: options.repository,
-        provider: options.provider,
-        settings,
-        credentialStore,
-      });
-      console.log(
-        `Configured ${name}: ${metadata.repository}, ${metadata.profile}, ${metadata.model}, ` +
-          `GitHub ${metadata.github.account} via gh, OpenRouter via OS keyring`
-      );
-    },
+    ...createTargetServices({ runtime, settings, httpTransport, requireTarget }),
 
     async capsuleCreate(options) {
       const context = await createSessionContext(options.target, runtime, settings);
